Version 1-18
GTKMM has been added to the list of programs which vimake can build without the need
of a Makefile. This allows the gtkmm demonstration package to be built as separate
stand alone programs. The GUI IDE has now been re-written as a vim plug-in, using gtkmm
programs to display toggle key messages. The tutorial has been updated to include all of
the new features.

Copy the vimake-1.18.tar.gz package to your home directory.
Then use the following commands to build and install.
      tar -zxvf vimake-1.18.tar.gz
      cd ~/vimake-1.18
      ./configure
      make install

Then change to the doc sub directory and run Tutorial.html This contains full Installation-
Configuration and a Tutorial on using Vimake in Vim, Gvim and the GUI.